refactored authentication
