+++ api/app/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI, APIRouter
 from fastapi.middleware.cors import CORSMiddleware
 import os
-from routers import birds, accounts, sightings, wish, family
+from routers import birds, accounts, sightings, wish, family, authentication
 from birddata import load_birds
 
 
@@ -29,7 +29,7 @@ app.add_middleware(
 
 app.include_router(birds.router, tags=['Birds'])
 app.include_router(accounts.router, tags=['Accounts'])
-# app.include_router(authenticator.router, tags=['Login / Logout'])
+app.include_router(authentication.router, tags=['Login / Logout'])
 app.include_router(sightings.router, tags=['Sightings'])
 app.include_router(wish.router, tags=['Wish List'])
 app.include_router(family.router, tags=['Family'])
